Revert to minimal changes to compile under Mac OS X and ran uncrustify

+++ VirtualH89/Src/HostFileBdos.cpp
@@ -181,8 +181,7 @@ int
 HostFileBdos::closeFile(BYTE* msgbuf, int len)
 {
     int         rc;
-
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (!fcb->s1[1])
@@ -213,7 +212,7 @@ HostFileBdos::closeFile(BYTE* msgbuf, int len)
 int
 HostFileBdos::searchNext(BYTE* msgbuf, int len)
 {
-    BYTE  u    = msgbuf[1] & 0x1f; // unused
+    BYTE  u    = msgbuf[1] & 0x1f;
     msgbuf[0] = 0;
     char* name = doSearch(&curSearch);
     if (!name)
@@ -290,7 +289,7 @@ HostFileBdos::deleteFile(BYTE* msgbuf, int len)
 int
 HostFileBdos::readSeq(BYTE* msgbuf, int len)
 {
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (fcb->fd <= 0)
@@ -305,7 +304,7 @@ HostFileBdos::readSeq(BYTE* msgbuf, int len)
         len *= 128;
         lseek(fcb->fd, len, SEEK_SET);
     }
-    long rc = read(fcb->fd, &msgbuf[0x25], 128);
+    int rc = read(fcb->fd, &msgbuf[0x25], 128);
     if (fcb->cr > 127)
     {
         fcb->cr   = 0;
@@ -333,7 +332,7 @@ HostFileBdos::readSeq(BYTE* msgbuf, int len)
 int
 HostFileBdos::writeSeq(BYTE* msgbuf, int len)
 {
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (fcb->fd <= 0)
@@ -341,7 +340,7 @@ HostFileBdos::writeSeq(BYTE* msgbuf, int len)
         msgbuf[0] = 9;
         return 1;
     }
-    long rc = write(fcb->fd, &msgbuf[0x25], 128);
+    int rc = write(fcb->fd, &msgbuf[0x25], 128);
     if (fcb->cr > 127)
     {
         fcb->cr   = 0;
@@ -453,7 +452,7 @@ HostFileBdos::renameFile(BYTE* msgbuf, int len)
 int
 HostFileBdos::readRand(BYTE* msgbuf, int len)
 {
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (fcb->fd <= 0)
@@ -462,7 +461,7 @@ HostFileBdos::readRand(BYTE* msgbuf, int len)
         return 1;
     }
     seekFile(fcb);
-    long rc = read(fcb->fd, &msgbuf[0x25], 128);
+    int rc = read(fcb->fd, &msgbuf[0x25], 128);
     seekFile(fcb);
     if (rc < 0)
     {
@@ -482,7 +481,7 @@ HostFileBdos::readRand(BYTE* msgbuf, int len)
 int
 HostFileBdos::writeRand(BYTE* msgbuf, int len)
 {
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (fcb->fd <= 0)
@@ -491,7 +490,7 @@ HostFileBdos::writeRand(BYTE* msgbuf, int len)
         return 1;
     }
     seekFile(fcb);
-    long rc = write(fcb->fd, &msgbuf[0x25], 128);
+    int rc = write(fcb->fd, &msgbuf[0x25], 128);
     seekFile(fcb);
     if (rc < 0)
     {
@@ -510,8 +509,7 @@ HostFileBdos::writeRand(BYTE* msgbuf, int len)
 int
 HostFileBdos::setRandRec(BYTE* msgbuf, int len)
 {
-
-    BYTE        u   = msgbuf[0] & 0x1f; // unused
+    BYTE        u   = msgbuf[0] & 0x1f;
     struct fcb* fcb = (struct fcb*) &msgbuf[1];
     msgbuf[0] = 0;
     if (fcb->fd <= 0)
@@ -632,7 +630,7 @@ HostFileBdos::getTime(BYTE* msgbuf, int len)
     localtime_r(&now, &tmv);
     // date is UTC, time is local, need to reconcile them...
     now      += tmv.tm_gmtoff;
-    long      date = now / 86400 - 2922 + 1;
+    int       date = now / 86400 - 2922 + 1;
     msgbuf[0] = (date & 0x0ff);
     msgbuf[1] = ((date >> 8) & 0x0ff);
     msgbuf[2] = ((tmv.tm_hour / 10) << 4) | (tmv.tm_hour % 10);
